ISSUE-345: split controller
